Remove argmax workaround decomposition
